[FIX] website: use user's language for configurator
The configurator introduced at [1] uses a frontend controller to display
its content. However this results in inconsistent language translation
as the python code uses the users's language but the localization
services uses the frontend's language when on a frontend page.

Steps to reproduce:
    - Have a website with a language different than the user's
    - Visit said website in the language different than the user's
    - Create a new website
    - The configurator's pages are in the website's language except
      for the features and industries

With this commit, if the user is in the website_publisher group, the
session will be populated with the user_context, giving the localization
services the user's language resulting in the configurator being
translated into just this language. This has a side effect however :
Some strings will be translated in the user's language while
in the frontend even if visiting another language. (e.g:
- editing an english website while user's language is french
- clicking on discord -> warning message is in french while the editor
is in english)

[1]: e8a5af2e28a4724f86783746706cb59175f086bf

task-2687416