[FIX] ProgressBar field issues
This commit fixes some bugs with the ProgressBarField recently added.
One bug occurs when max_value and/or current_value are given
directly instead of being a data from the record. The other issue is
related to the values not being resetted after changes have been discarted.
It is now resetting correctly