[ADD] payment_stripe_connect_proxy: integrate Stripe Onboarding
This commit intergrate the Stripe Onboarding through payment Onboarding
using the iap proxy.

Purpose
=======
Help users to easily onboard with Stripe by using the Stripe Connect
API.

Specifications
==============
During the payment onboarding, if the user selects Stripe, the user
will start the Stripe Onboarding :
NB : The process uses a proxy which handles the Stripe Onboarding calls
and sign them with the Stripe Connect key.
1) A call is made through the proxy to get the Stripe account token;
2) A call is made through the proxy to get the Stripe account link,
which contains the url of the Onboarding;
3) The user is redirected to the Stripe Onboarding;
4) The user completes the Stripe Onboarding;
4) The user comes back to Stripe Acquirer Form.

Note that the Onboarding status isn't stored in the database so there
isn't a call to Stripe API to validate the account status.

API Documentation : https://stripe.com/docs/connect/standard-accounts

task-2685160