[IMP] payment_stripe_connect_saas: configure webhook
This commit allows to directly configure webhook on finishing Stripe
Onboarding.

I have strong concerns about this feature as it creates webhook on the
(master) Connect account and so every instance using a Stripe account
connected to the (master) Connect account will receive every trigger
attached to any other connected account.

I fear we will have lots of webhooks defined on the (master) Connect
account which will have lots of triggers to send on lots of instances.

The recommandation is to create a single webhook for every connected
account and then, on the Platform side, posting the event to the right
instance.

task-2685160