[ADD] web: add odoo ui icons
This commit will add a custom icons webfont meant to coexist with
FontAwesome 4.7 and to be extended over time.

The goal is to provide a wider range of icons specifically made for our
UI and to cover FontAwesome's lack of iconography for specific business
and functionality.

Class names will be available in our documentation website and the
source SVGs files can be found in our design-assets repo.

This commit is part of v16 overall restyle, task#2704984.