[IMP] mrp: update WO state when all previous WOs cancelled
In case of backorder, it can happen that the first WO is already done in
the original MO, so it get cancelled in the backorder MO. In that case,
the second WO is actually the first WO available. In this case, we will
make the second WO's state "Ready" or "Wait". Similarly, if all previous
WOs cancelled, the first one available should be updated.

Task-2678388