[IMP] l10n_in: added PAN in partner
Permanent Account Number (PAN) is a ten-digit alphanumeric number, issued in the
form of a laminated card, by the Income Tax Department, to any "person" who
applies for it or to whom the department allots the number without an application.

PAN enables the department to link all transactions of the "person" with the
department. These transactions include tax payments, TDS/TCS credits, returns of
income, specified transactions, correspondence, and so on. PAN, thus, acts as
an identifier for the "person" with the tax department.

reference - https://bit.ly/3sFKI8A

Task-2610237
PR-#75613