[IMP] website, *: handle url redirections from website iframe
*: portal, auth_signup, website_knowledge

When navigating in the iframe, only same origin redirections should be
open in the iframe contentWindow. External redirections should be done
in the top window.

Some internal pages had to be served with X-Frame-Options header set to
SAMEORIGIN, and Content-Security-Policy to "frame-ancestors 'self'" (see
[1]).

All the links that are redirecting to another host, and the client
actions, are opened in the top window.
Exemples that will be opened in the iframe's top window:
- Clicking on a link google.be that should not open in a new tab
- Clicking on the language selector and adding a new one, or
clicking on "logout".

For that, an overridable _isTopWindow method is introduced on the client
action, which will also be used to redirect when a wrong path param is
given to the client action.

task-2687506

Co-authored-by: Arthur Detroux <ard@odoo.com>