[IMP] website, *: handle url redirections from website iframe
*: portal, auth_signup

When navigating in the iframe, only same origin redirections should be
open in the iframe contentWindow. Some internal pages had to be served
with X-Frame-Options header set to SAMEORIGIN.

Also, all the links that are redirecting to another host, and the client
actions, are opened in the top window.
Exemples that will be opened in the iframe's top window:
- Clicking on a link google.be that should not open in a new tab
- Clicking on the language selector and adding a new one, or
clicking on "logout".

task-2687506