[FIX] point_of_sale: infinite loop update-render due to get_unit_display_price
This is a temporary fix.
The orderline component was calling a getter to get the unit display price. The getter updates (that shouldn't happen) the quantity of the orderline temporarily in order to call `get_all_prices()` (which use the quantity of the orderline). Due to the update of an attribute, this was triggering the render call of the component. => Infinite loop