[FIX] website: fix infinite loading after creating a website
The website configurator introduced at [1] redirects the user after
choosing a theme.
Prior to this commit a user could get stuck in a infinite loading
screen if they tried creating a website while having previously
visited the front-end in a different language than the site
they're creating.

Steps to reproduce :

- Log into a database and have more than one language on a Website
- Visit said website in a different language than the default one
- Create a new website
- Go through the configurator and click on Build my website
- Once on the website, the loading animation doesn't disappear.

The reason behind this is that the configurator is automatically started
when visiting the page and is supposed to remove the loading animation.
However when visiting the website in another language, the editor code
isn't loaded and therefore, doesn't remove the loading animation.
With this commit, we make sure that the user is always redirected to
the default language of their new website when redirecting with
?enable_editor=1 so that it can initialize properly.

[1]: e8a5af2e28a4724f86783746706cb59175f086bf

task-2687416