[IMP] website_sale: add tour for snippet products
This PR added new templates and features to the dynamic snippet
products. The goal of this commit is to test these features by selecting
every newly made template, saving, and adding a product to the cart.

Another tour was added to test the functionality of the recently
viewed product filter.

task-2677203