[IMP] stock: add not-linked MTO option to rules
Currently for a stock.rule, when:
- action = pull and procure_method == make_to_order
- action = push  and auto == manual
Moves are auto-generated and auto-linked (i.e. move_dest_ids are set).
Sometimes it's desired that the moves are auto-generated, but
not auto-linked Therefore we add in this option to stock.rule and by
default it is True so we follow the existing behavior by default.

There is a little hack to make this work correctly in the MO pull case
(i.e. make sure stock.move._adjust_procure_method() does not bypass this
rule setting) where we assign the pull rule being applied to the
move.rule_id.

Task: 2680470