[IMP] knowledge, web, mail: search articles with command palette
Allow users to find Knowledge articles with the command palette,
which can be opened with CTRL+K or anywhere the palette can
be opened.

Articles can be searched by typing "?" followed by the query string.
The query string is used to match the name of the article and the
article's parent name (content of the article is not used).

Once the user chose an article, they will be redirected to the selected
article in the knowledge app. Later commits will allow more
integrations between the knowledge app and chatter in the other apps,
to include, notably, some parts of an article in notes or messages on
other models.

A bridge between legacy and owl has been made in
web/static/src/legacy/utils.js since form views don't yet support owl.
The function splitCommandName in web/static/src/core/commands/command_palette.js
was exported to be used to highlight parent name as well.

Open palette on chatter and article view
----------------------------------------

Show a button on any chatter that opens the command palette with "?" as
a placeholder to quickly search for knowledge articles.

The palette is also opened with "?" when clicking on the search bar in
knowledge front view.

Task-2715436