[FIX] sale: remove dumb onchange logic
Do not update the commitment_date everytime the expected date is
modified, otherwise there is no meaning to have two fields...

Furthermore, it seems that this onchange causes a performance loss
on SO/SOL updates, forcing the whole SOL data to be returned along
the SO data during the SO onchange.

Therefore removing this onchange improves the performance of SOL updates
and consequently the final save of the record by reducing the data wrongly
considered modified and given to the create/write call.

X-original-commit: 116ca30b2f5f27f8959961c786b88fdcec36338a