[TO SQUASH] rework model for acces rules and membership
The access are defined by the two following fields:
internal_permission = selection (none / read / write)
article_member_ids = decorated m2m
    - article_id (mandatory)
    - partner_id (mandatory)
    - permission = selection field (none / read / write)

The category in which an article is depends on both fields.
- Workspace: internal_permission != none. Everyone can see it.
    There can also be some members on the article.
    Note: Everyone = every internal users.
    E.G.: Everyone can see it except Marc Demo. (member.permission = 'none')
          Everyone can read it but Only Marc Demo can write.
          Everyone can see it + some external partners.
- Shared: internal_permission == none + have multiple members.
    Only the members can see it.
- Private: internal_permission == none + Only one member AND have 'write' access
    This member is the 'owner' and only them can see it.

An admin can always see and edit every article.

The current user (except for admin) has no access to articles:
- that are not shared with them:
    internal_permission = 'none' and not a member or member.permission='none'
- that are private but is not the owner:
    internal_permission = 'none' and not the owner
- that are 'public' (in workspace) but user have specifically
member.permission = 'none'

The current user (except for admin) can edit an article only if:
- internal_permission = 'write' AND not member.permission = none
OR
- is member and member.permission='write'

Editing an article means that the user can change write on the article but also
change the access rules of the article. But, a user that can only see an article
can only share the link of the article, and consequently, add a member but in
read access only.

Constraint:
An article should always have at least one user having write access on it.
In other words, there can't be any article in the following situations:
- internal_permission = 'none' and no members
- internal_permission = 'none' and only one member that have read access
- internal_permission = 'read' and no members with write access.

Access and membership heritage: