[IMP] web_editor: add media dialog
The media dialog was refactored using owl.
It can be displayed via the dialogs service and uses following props:
- resModel
- resId
- onlyImages
- noImages
- noDocuments
- noIcons
- noVideos
- multiImages
- useMediaLibrary
- media
- save

The (Image/Document/Icon/Video)Selector components are controlled by the
state of their parent, the MediaDialog: this.selectedMedia.
Unfortunately, it looks like that the whole state of each tab has to be
passed to the child Selector, otherwise the UI is not correctly updated.

Some things differ from the original implementation: the dialog still
takes (optionally) a media as a parameter, but it does not directly
modify it anymore.
It will pass the selected media as a parameter to a callback, save, that
it gets as a prop. That save prop should be responsible to make the
appropriate modification to the original media, given the new selected
media.
Some adapations will probably be needed when using it to the upcoming
wysiwyg adapter component.

There are still some implementation details that should be improved,
the code can be confusing. It opens up for an easy multiSelect of other
tabs than the images (multiIcons, multiDocuments), but that use case is
not currently tackled and maybe it would be better to simplify the code.

task-2687506