[IMP] website: remove unused assets
This commit removes the website navbar root widget, as the edition is
now handled in the backend through the client action and custom menus.

It removes associated widgets (the fullscreen indication, the page
properties dialogs, ...) and their css.

The o-w-preserve-XXX are still left, as they are used for the technical
modals the website_slides is using.

task-2687506