[FIX] mrp: backport fix cancel backorder
Backport of the following fix: c494ba2b8b40e939a311188e5433857e299bdfe6
Sadly, because of the creation of a new stored field, the all commit
can't be fully backport, and so, only the part about the workorders
cancellation is backported (the part about the propagation of the
quantity to compute more precisely the remaining quantity is aborted).

Also, adds a condition to avoid to take canceled workorders to compute
the remaining quantity, so it's avoid to recreate a new workorder for
previously canceled workorder.

Task: 2751230