[IMP] web: remove LegacyComponent on HotkeyCommandItem
With owl2, the trigger of components is no longer available. It still
works in Odoo on LegacyComponent, which has been introduced to ease
the switch from owl1 to owl2, and we're now incrementally removing
it.

This commit replaces the use of `trigger` on HotkeyCommandItem
by a prop and sets Component as the base class.