[IMP] web: remove LegacyComponent on PivotRenderer
With owl2, the el of components is no longer available. It still
works in Odoo on LegacyComponent, which has been introduced to ease
the switch from owl1 to owl2, and we're now incrementally removing
it.

This commit replaces the use of `el` by a ref on PivotRenderer and
sets Component as the base class.