[FIX] website: properly catch click events.
Prior to this commit, clicking on a link on website could change page
even in edit mode. This isn't ok as we do not want the browser to change
page while editing the website, leading to possible loss of changes.
And mostly crashing the editor.

This commit prevent the default action of clicking on elements when the
editor is active.