[IMP] website: improve management of edition state
Prior to this commit, the edition could start before jquery was even
loaded. This would cause a traceback and justifiably so.
The issue is that waiting for it to load is now an issue as it slows
down everything quite a bit.

For this to be fixed we need to improve the things that are loaded by
the editor_helper.