[IMP] website: review of editor states
This commit is a review of the various states introduced on the
editor component. It adds a new event from the iframe to let the
context know that the PublicRootWidget inside the iframe is ready.

It also fixes the save feature : Previously, an event was added on the
button to trigger a saveContent method on the wysiwyg widget. This was
an error as the SnippetsMenu is already listening on the save button to
trigger a save request event.