[IMP] website: cache snippets template at the module level
Move the cache system of the snippets template at the module level
so that it survives different SnippetsMenu instance.

This wasn't necessary before because if the SnippetsMenu is killed,
the entire page was reloaded. Now that the JS Modules are kept through
different instance of the SnippetsMenu, it's interesting to keep the
template on a cache that's at the higher level to improve performance.

Co-authored-by: Younn Olivier <yol@odoo.com>