[IMP] web_editor: add navigation node
Since /template and /file will be UNREMOVABLE and CONTENTEDITABLE=FALSE, it may
be difficult to add a text node between such blocks once they are siblings.

This new concept uses a temporary `<P>` node that will appears in between blocks
or while exiting a block using ArrowUp and ArrowDown keys. This node is
temporary and will disappear if the selection change. If the user starts typing
in such a block, it will be validated in the dom and wont be removed anymore.

Such a node will also be useful with existing blocks (table, code, ...)