[IMP] web_editor: remove last empty <p>
Since the navigation node may add unwanted `<P>` in places where it is not
currently possible to remove them, (i.e. between two /template block), add the
possibility to force remove them with CTRL(/META) + BACKSPACE, with the
condition that it is not currently the last element in a contentEditable block.

This will also prevents the user to remove the last empty `<P>` of a field_html,
which is a good thing.