[IMP] website, base: adapt reload bundles to work with the new editor
Prior to this commit, change the style of the website would apply the
changes within the iframe but the editor wouldn't reflect the new values
(e.g colors wouldn't be updated).

This commit reviews the reloadBundles method so that it applies the
changes to the webclient as well. It also adapts the routes
to return the right debug packages. Effectively removing the need to
reload the page at any time.