[FIX] web_editor: historyRevert a remove from parent
The last mutation registered in case of a remove from parent with multiple
children is a remove of the last child.

To undo such a remove, we will go through the mutations in reverse order, and as
such, the first children will be added last, so the operation to do is prepend
and not append.