[IMP] web_editor: in-depth restore of unremovables (deleteRange)
Currently, unremovables are being restored but their textContent is forcibly
reinitialized.

Since Knowledge introduces new Unremovable Blocks, which are containers with
interactive buttons and subelements which constitute a structurally consistent
entity, we want to disallow removing and partially restoring such blocks during
a deleteRange.

The proposed solution is to use a combination of the fact that a block is
UNREMOVABLE and its CONTENTEDITABLE attribute to determine if its content needs
to be reinitialized or not:

An UNREMOVABLE CONTENTEDITABLE=false will keep its children, and an UNREMOVABLE
CONTENTEDITABLE=true will lose them. UNREMOVABLE subchildren of another
UNREMOVABLE are also evaluated with the same logic, except that an
sub-UNREMOVABLE CONTENTEDITABLE=true will not reset its content if it is a
descendant of an UNREMOVABLE CONTENTEDITABLE=false

In the /template KNOWLEDGE case, we have the root block which is UNREMOVABLE
CONTENTEDITABLE=false which will be restored in its entirety, along with its
editable content which is UNREMOVABLE CONTENTEDITABLE=true -> content is
preserved, since descendant of an UNREMOVABLE CONTENTEDITABLE=false.