[FIX] website: adapt web_editor patching and extends
Prior to this commit, the web_editor javascript modules were patched.
This didn't cause any issue as the web_editor for website, and the
web_editor for the backend where completely distinct and never existed
at the same time.

Now that the website wysiwyg editor also exists in the backend,
it means that the website modules can be loaded at the same time
as the mass_mailing modules. This revealed some bugs as bits of code
didn't work in that situation. Mainly, a `widgets_start_request` event
that never got a callback.

To fix this issue, this commit tries to extend modules as much as
possible instead of extending them and, where it's not possible to
extend, it calls method conditionally by checking if it's in a website
edition environnement. This should be the norm from now on.