[IMP] web_editor: add failsafe upon CTRL+BACKSPACE
Prevents the user to remove the only <p> child of a contenteditable=true, in
order to prevent the user from adding text nodes directly (outside of a <p>
element), which was possible before (removing the only <p> element of a
contenteditable=true would then allow the user to place the carret in the
editable and type, which would create a text node outside of a <p> element.

Add a failsafe on CTRL+BACKSPACE that will inspect the children of the parent
after the CTRL+BACKSPACE took effect, and add an empty <p><br></p> if there is
no child left in the parent. It will reduce occurences of having the caret in
an element that is not supposed to have text nodes (i.e. this.editable).

Task-ID: 2794990