[ADD] tools: add XML files formatting function
When generating XML files in Odoo from a template (with a _render function for instance), the result is often not well formatted. There are some blank spaces and XML tags find themselves on wrong lines. This leads to several modules having to implement similar ways of indenting/formatting their XML files.

This commit adds a function in tools that can be used to circumvent such problems and to provide a pretty formatting for XML files.