[FIX] website: use correct endpoint in get_canonical_url
`router` is a routing map which might be different than the request's.
In this case, `request.endpoint` won't be found as a key of this routing
map, and it will lead to a werkzeug.routing.Builderror.

By re-matching the current path on this specific router, we make sure
that the endpoint will be found in this routing map.