TO DELETE [IMP] web_editor: master editor knowledge abd
TO DELETE [FIX] web_editor: prevent deleteBackspace to move nodes

Before this commit, whenever the element that must be deleted by
`deleteBackspace` matches `isMediaElement` or `isNotEditableNode` and
the cursor was in the beginning of a `<p>` with inline content, the
inline content moved to the parent and the `<p>` was deleted.
Which is wrong.

Now, whenever the element that must be deleted by `deleteBackspace`
matches `isMediaElement` or `isNotEditableNode`, no inline content
moves.

Task-2781328

TO DELETE [FIX] web_editor: make HR element not editable

task-2778511

[FIX] web_editor: isVisibleEmpty (with /image empty <a>)

If node is an empty <a> which has o_image, like what is constructed when using
/image powerbox command, isVisibleEmpty should return true

Task-ID: 2794990

[FIX] web_editor: historyRevert a remove from parent

The last mutation registered in case of a remove from parent with multiple
children is a remove of the last child.

To undo such a remove, we will go through the mutations in reverse order, and as
such, the first children will be added last, so the operation to do is prepend
and not append.

Task-ID: 2794990

[IMP] web_editor: add navigation node

Since new commands for the Knowledge module (like /template and /file) will be
CONTENTEDITABLE=FALSE, it may be difficult to add a text node between such
blocks once they are siblings.

This new concept uses a temporary `<P>` node that will appear in between blocks
or while exiting a block using ArrowUp and ArrowDown keys. This node is
temporary and will disappear if the selection change. If the user starts typing
in such a block, it will be validated in the dom and won't be removed anymore.

Such a node will also be useful with existing blocks (table, code, ...)

Task-ID: 2794990

[IMP] web_editor: add failsafe upon CTRL+BACKSPACE

Prevents the user to remove the only <p> child of a contenteditable=true, in
order to prevent the user from adding text nodes directly (outside of a <p>
element), which was possible before (removing the only <p> element of a
contenteditable=true would then allow the user to place the carret in the
editable and type, which would create a text node outside of a <p> element.

Add a failsafe on CTRL+BACKSPACE that will inspect the children of the parent
after the CTRL+BACKSPACE took effect, and add an empty <p><br></p> if there is
no child left in the parent. It will reduce occurences of having the caret in
an element that is not supposed to have text nodes (i.e. this.editable).

Task-ID: 2794990

[IMP] web_editor: prevent command refocus when unnecessary

Currently, the focus is forced on this.editable when using a powerbox command.

Knowledge /template will introduce blocks which are contentEditable=False, but
contains sub-elements with contentEditable=True, which are allowed to use
powerbox commands. Those sub-elements have another editable context as
this.editable. As such, forcing a refocus on this.editable will remove the
caret from the sub-element.

Avoid doing the refocus if the currently focused element is a child of
this.editable and is contentEditable=true.

Task-ID: 2794990

[IMP] web_editor: flag only editable as unremovable

Prior to this change, if a contenteditable=true element was a child of a
contenteditable=false element, it was always flagged as unremovable=true.

In fact what should happen is that this.editable should be unremovable, but one
of its subchildren can be contenteditable=true while having a parent
contenteditable=false and not be unremovable.

Task-ID: 2794990

[IMP] web_editor: expand deleterange for uneditables

In Chrome, it is possible to partially select the contents of a
contenteditable=false element if it has a contenteditable=true child.

This commit prevents breaking such uneditables by expanding the range in a
deleterange to fully contain such elements.

It will be useful with the new blocks introduced with the Knowledge module.

Task-ID: 2794990

[IMP] web_editor: add empty <P> on click

Insert an empty <p><br></p> when clicking below the last child of this.editable
if it is not already a <p> element so that even if this.editable happens to be
empty or only contains contenteditable=false elements, the user is able to type
and place the caret somewhere. It will also be more user friendly when the last
element is contenteditable=false.

Implement various failsafes in order to add a <p> element to handle the caret if
the user clicks on an element that should be able to handle the caret but
currently can not or incorrectly (i.e. will add text nodes instead of <p>
elements).

1) When clicking on an editableContextParent which only has uneditable children,
append a navigationNode

2) When clicking on an editableContextParent which has no child, append a
<p><br></p> element

(an editableContextParent is an element which is contenteditable=true contained
in an element contenteditable=false)

Task-ID: 2794990

[IMP] web_editor: add uneditable selection for oDeleteBackward

> Preamble

The first objective of this commit is to be able to remove a block in between 2
contenteditable="false".

It is also a POC for a more complete behavior which would handle every
OdooEditor blocks so that they can be selected on click, with arrow keys, ...
Such a behavior could replace the navigationNodes.

A future improvement could be to change the color of the selection in
collaborative (use the same color as the caret).

> Content

Add the attribute 'contenteditable' to the frozen list for html elements of
DOMPurify, so that this attribute is not sanitized in collaborative mode when
using _safeSetAttribute.

Allows to select a contenteditable=false element using oDeleteBackward.
This fake selection will replace the normal document selection.
While the fake selection is active, the user can delete the selected element
with 'Backspace' or 'Delete', write a character which will replace it, use the
arrows to navigate from it, or change the selection by any means to replace the
fake one by a real one.

Task-ID: 2794990

[IMP] web_editor: use the uneditable selection for <table>

Allows <table> elements to benefit from the uneditableSelection behavior for
deletion.

Task-ID: 2794990