[FIX] Web_editor: fix element merge on deleteBackward
Ensure we never move nodes that were not initially in the deleted element.

When Deleting an element and merging it's content with is sibling,
the nodes contained in the removed element will temporary be hosted
in the common parent, beside possible other nodes.

We don't ever want to touch those others node when merging.
This change fix some instances where nodes would wrongly be carry
into the another element.