[IMP] website: add cache for _get_canonical_url_localized
Optimize cache of url_rewrite, qs is ignored and only useful for
RequestRedirect exception of werkzeug that we already handle manually.

This method is called at least once by lang by request. On a website of ~20
langs (e.g. on a staging of odoo.com) it improves the page speed from:
79 0.076 0.095
  to
45 0.028 0.056

On a runbot with 5 langs installed, we can see a small improvement from
61 0.029 0.034
  to
49 0.023 0.030

to render a blog post.