[FIX] core: recompute fields depending on relational computed fields
The use-case is a dependency chain of fields: field C depends on the
inverse of relational field B, which itself depends on field A.  When
field A is modified, field C should be recomputed on the records that
indirectly depend on the former value of A.

Indeed, consider the situation below, where the value of field B is
represented as the down arrow.  The left and right diagrams represent
the relation before and after the update of A, respectively.

        line(A, B)                                 line(A, B)
            |                                          |
            v                                          v
        record1(C) record2(C)           record1(C) record2(C)

If the relation B changes, the field C should be recomputed on both
records, even if field A is not relational per se.  From the point of
view of recomputation triggers, field A should be considered as a
relational field, and the framework must trigger the computation of B
and C on the records before and after the update of A.