[IMP] website: add website navbar AceEditor
The goal of this commit is to adapt the website HTML/CSS/JS editor using
a 'ComponentAdapter' to website backend.

Since we reload the iframe only after editing resources, the '_updateEditor'
method was added to update UI (toggle 'resetButton' visibility,...) and get
the same behaviour (on save/reset) as before using iframe.

task-2687506