[IMP] website: add website navbar AceEditor
The goal of this commit is to adapt the website HTML/CSS/JS editor using
a ComponentAdapter to website backend.

fixup: Since we reload the iframe only after editing resources on
aceEditor, We need to refresh editor's UI after the update.

The goal of this commit is to add '_updateEditor' method to get the same
behaviour (on save/reset) as before using iframe.

task-2687506