[IMP] website, web_editor: add WebsiteEditor components
This commit adds 2 components essential to website edition.

- WysiwygAdapterComponent: Allows the legacy @web_editor/wysiwyg widget
    to work in an OWL context. It will be further expanded in the next
    commit to fully replace the previous edit.js wwhich was the initial
    widget that started the edition prior to this commit / branch.

- WebsiteEditorComponent: This component is in between WebsitePreview and
    the adapter. It allows some tricks described below.

The reason this is split in two component is because, for now, when
the page needs reloading, the editor component will clone the content
of the widget, destroy the Wysiwyg, and restart it. This ensure that every
event is properly bound and gives the illusion that the wysiwyg is
never destroyed.

task-2687506