[FIX] web_editor: avoid instantiation of tooltip if not already done
In BS5 we can't add two components' instance on an element.
So before disposing a tooltip we check if there is an instance of a
tooltip.

e.g.:
> Bootstrap doesn't allow more than one instance per element. Bound
> instance: bs.carousel.

[FIX] web: avoid initialize tooltip when no need

As BS5 don't allow to use multiple component on the same node, we avoid
useless tooltip as in the Pivot view there is "Measures" button that is
a dropdown. So it generates an error due to the pre-initialisation of
the tooltip and then the dropdown on the button.

> Bootstrap doesn't allow more than one instance per element.
> Bound instance: bs.tooltip.